requirements.txt
pynput>=1.7.6
obsws-python>=1.7.0
pythonosc>=1.8.0
pyyaml>=6.0
